Disable loading fixture data
The commands to load fixture data for industries, index, about, and
testimonials have been commented out. This prevents the loading of potentially
outdated or unnecessary data when running the entrypoint script.

The fixtures can be loaded manually if needed, but they will no longer be
automatically loaded on every run of the production entrypoint script.